Improved entity resolution search
Queries refactored so that each entity row includes all phone numbers
associated with that entity, instead of only the phone number in the
search query.

Uses nested postgres queries, which is kind of an antipattern. Though it
works, we need to replace the callbacks with promises at some point to
maintain a more linear structure for the app.

Fix #3